% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnival_helpers.R
\name{assignPROGENyScores}
\alias{assignPROGENyScores}
\title{assigny progeny scores to genes via OmnipathR}
\usage{
assignPROGENyScores(
  progeny = progeny,
  progenyMembers = progenyMembers,
  id = "gene",
  access_idx = 1
)
}
\arguments{
\item{progeny}{table of Progeny output}

\item{progenyMembers}{List of Members from Progeny}

\item{id}{ID of gnees}

\item{access_idx}{Score Access ID}
}
\description{
assigny progeny scores to genes via OmnipathR
}
\examples{
NULL
}
