% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_functions.R
\name{better_dotplot}
\alias{better_dotplot}
\title{improved code to perform a dotplot}
\usage{
better_dotplot(gset, c_groups = contrast_groups)
}
\arguments{
\item{gset}{GSEA result from Cluster Profiler}

\item{c_groups}{contrast groups vector of length 2.}
}
\value{
list of length two with two dotplot reults
}
\description{
improved code to perform a dotplot
}
\examples{
NULL
}
