% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{check_against_index}
\alias{check_against_index}
\title{test}
\usage{
check_against_index(info, index)
}
\arguments{
\item{info}{Info field to check}

\item{index}{index that is to be checked}
}
\value{

}
\description{
test
}
\examples{
NULL
}
