% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnival_helpers.R
\name{convertHumanGeneHomologs}
\alias{convertHumanGeneHomologs}
\title{convert between human and mouse homologs using homolog table}
\usage{
convertHumanGeneHomologs(
  x,
  jax_database = NULL,
  input_organism = "Homo sapiens"
)
}
\arguments{
\item{x}{List of genes}

\item{jax_database}{Database from outside to reduce http calls}

\item{input_organism}{Organism being converted from (Homo sapiens or Mus musculus)}
}
\description{
convert between human and mouse homologs using homolog table
}
\examples{
convertHumanGeneHomologs(c("APLNR", "APLNR"))
}
