% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnival_helpers.R
\name{convertHumanGeneList}
\alias{convertHumanGeneList}
\title{Convert Human gene list}
\usage{
convertHumanGeneList(x, hmart, mmart)
}
\arguments{
\item{x}{character vektor of gene symbols to translate (DEPrecated)}

\item{hmart}{hunan biomart}

\item{mmart}{mouse biomart}
}
\description{
Convert Human gene list
}
\examples{
NULL
}
