% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{create_overlap_matrix}
\alias{create_overlap_matrix}
\title{Create dynamic table from significant gene names}
\usage{
create_overlap_matrix(sig_gene_names)
}
\arguments{
\item{sig_gene_names}{import list of genes to put in matrix}
}
\value{
DT::datatable of genes across comparison groups
}
\description{
Create dynamic table from significant gene names
}
