% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnival_helpers.R
\name{df_to_viper_regulon}
\alias{df_to_viper_regulon}
\title{\code{df_to_viper_regulon}}
\usage{
df_to_viper_regulon(df)
}
\arguments{
\item{df}{a dataframe of n*3 dimension. The first column corresponds the targets,
and the second column indicates which regulon does each target belongs to.
The third column corresponds to the weight and sign of the interaction between
a regulon and its targets.}
}
\value{
a list where each element is a regulon in the viper format. This list
is ready to be used as a regulon set in viper.
}
\description{
This function is designed to generate a ready to use regulon object for viper
from a 3 column dataframe representation of a target set collection.
}
