% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{filter_split_table}
\alias{filter_split_table}
\title{filters table so only contrasted samples remain}
\usage{
filter_split_table(rld, contrast_groups, j_df, reorder_genes = TRUE, cond_id)
}
\arguments{
\item{rld}{DESeq2 object that has been normalized with either (rlog2 or vst)}

\item{contrast_groups}{character vecctor of contrast groups being analyzed.}

\item{j_df}{joined tibble of deseq + fpkm values to match genes and gnames}

\item{reorder_genes}{whether the order of genes should be orded by logfolgchange}

\item{cond_id}{condition variable set in snakeamek}
}
\value{
matrix of epxression values
}
\description{
filters table so only contrasted samples remain
}
