% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{fpkm_to_tpm}
\alias{fpkm_to_tpm}
\title{Convert fpkm values to tpm}
\usage{
fpkm_to_tpm(fpkm_mat)
}
\arguments{
\item{fpkm_mat}{matrix of fpkm values (rows = genes, columns = samples)}
}
\value{
a matrix of tpm values
}
\description{
Convert fpkm values to tpm
}
