% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnival_helpers.R
\name{generateTFList}
\alias{generateTFList}
\title{Generate TF list from Saezlab Transcript tutorial}
\usage{
generateTFList(df = df, top = 50, access_idx = 1)
}
\arguments{
\item{df}{Dataframe from Dorothea}

\item{top}{Top n genes to}

\item{access_idx}{Col to access}
}
\value{

}
\description{
Generate TF list from Saezlab Transcript tutorial
}
\examples{
NULL
}
