% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_functions.R
\name{get_kegg_name}
\alias{get_kegg_name}
\title{Return the correct kegg name}
\usage{
get_kegg_name(org_name)
}
\arguments{
\item{org_name}{Name of organism used in analysis}
}
\description{
Return the correct kegg name
}
\examples{
NULL
}
