% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MitoCarta_mouse.R
\name{get_MitoCarta_mouse}
\alias{get_MitoCarta_mouse}
\title{Get MitoCarta Mouse Data}
\usage{
get_MitoCarta_mouse()
}
\value{
A data frame containing the processed MitoCarta mouse data.
}
\description{
This function downloads the MitoCarta mouse data from a specified URL, processes the data, and saves it to a local file.
It also returns the processed data.
}
\examples{
\dontrun{
get_MitoCarta_mouse()
}
}
