% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_functions.R
\name{get_org_db}
\alias{get_org_db}
\title{Return the correct org_db}
\usage{
get_org_db(org_name)
}
\arguments{
\item{org_name}{Name of organism used in analysis}
}
\value{

}
\description{
Return the correct org_db
}
\examples{
NULL
}
