% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnival_helpers.R
\name{get_organism_omnipath_name}
\alias{get_organism_omnipath_name}
\title{get organism name used in omnipath for Mus musculus or Homo sapiens}
\usage{
get_organism_omnipath_name(organism_name)
}
\arguments{
\item{organism_name}{Name of organism (Mus musculus, Homo sapiens)}
}
\description{
get organism name used in omnipath for Mus musculus or Homo sapiens
}
\examples{
NULL
}
