% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_functions.R
\name{gsea_test}
\alias{gsea_test}
\title{Run gene set enrichment analysis using \link[fgsea]{fgsea}}
\usage{
gsea_test(DE_genes, T2G, input_type = "gene_symbol", ...)
}
\arguments{
\item{DE_genes}{Tibble of diferentially expressed genes and their log fold
changes}

\item{T2G}{Term to Gene table of gene set enrichment.}

\item{input_type}{Gene ID type used in database}

\item{...}{Arguments passed to \link[clusterProfiler]{GSEA}}
}
\value{
enrichment result from \link[fgsea]{fgsea}
}
\description{
Run gene set enrichment analysis using \link[fgsea]{fgsea}
}
