% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{join_tables}
\alias{join_tables}
\title{Joins the DEseq table and the fpkm read table together}
\usage{
join_tables(DEseq_tb, fpkm)
}
\arguments{
\item{DEseq_tb}{Deseq result table from DEseq. Saved as TSV table by deseq rule}

\item{fpkm}{FPKM table from snakemake pipeline. You can use any fpkm table.}
}
\value{
sorted combined tibble between the input files.
}
\description{
Joins the DEseq table and the fpkm read table together
}
