% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{mean_tibble_from_mat}
\alias{mean_tibble_from_mat}
\title{Calculates mean values from the Four groups, which are still hardcoded. and returns
the mean values plus gene names and their adjusted p values}
\usage{
mean_tibble_from_mat(
  mat,
  extra_col = NULL,
  contrast_groups = contrast_groups,
  s_map
)
}
\arguments{
\item{mat}{Object that can be safely coerced to tibble with col names from Vascaging}

\item{extra_col}{Extra column that should be extracted and added}

\item{contrast_groups}{contrast_groups}

\item{s_map}{Sample map of samples}
}
\value{
tibble with mean values, gene names and padj values
}
\description{
Calculates mean values from the Four groups, which are still hardcoded. and returns
the mean values plus gene names and their adjusted p values
}
