% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mito_gs.R
\docType{data}
\name{MitoPathways}
\alias{MitoPathways}
\title{Mitochondiral gene sets from MitoPathways}
\format{
A Dataframe with x rows and y columns:
\describe{
 \item{Symbol}{Name of the gene as a gene symbol}
 \item{EnsemblGeneID}{Ensembl Gene ID}
 \item{GeneSet}{Name of the gene set the gene in target belongs to}
}
}
\usage{
MitoPathways
}
\description{
This object contains the gene sets for mitochondrial genes from MitoPathways.
}
\keyword{datasets}
