% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_functions.R
\name{over_rep_test}
\alias{over_rep_test}
\title{Run over representation}
\usage{
over_rep_test(DE_genes, T2G, universe, input_type = "ENSEMBL", ...)
}
\arguments{
\item{DE_genes}{Tibble of diferentially expressed genes and their log fold
changes}

\item{T2G}{Term to Gene table of gene set enrichment.}

\item{universe}{Universe of all tested genes from the experiment}

\item{input_type}{Gene input type used f.e. ENSEMBL or gene symbol}

\item{...}{passed to enricher function \link[clusterProfiler]{enricher}}
}
\value{
enrichResult object from \link[DOSE]{DOSE-package}
}
\description{
Run over representation
}
