% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_functions.R
\name{plot_enrichment}
\alias{plot_enrichment}
\title{Plot Enrichment Barplot}
\usage{
plot_enrichment(
  GSEA_table,
  X,
  Y,
  pval = "pval",
  pval_threshold = 0.1,
  n_max = 20
)
}
\arguments{
\item{GSEA_table}{Table containing results from GSEA analysis}

\item{X}{Column with Pathways names of GSEA}

\item{Y}{Column with NES/ES from GSEA}

\item{pval}{Column containing pvalue}

\item{pval_threshold}{pvalue filter threshold}

\item{n_max}{Max number of pathways to plot}
}
\description{
Plot Enrichment Barplot
}
\examples{
NULL
}
