% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnival_helpers.R
\name{process_carnival}
\alias{process_carnival}
\title{Parse and change variable types for carnival}
\usage{
process_carnival(carnival_res)
}
\arguments{
\item{carnival_res}{Carnival Results object}
}
\value{

}
\description{
Parse and change variable types for carnival
}
