% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{read_deseq_tibble}
\alias{read_deseq_tibble}
\title{Parse deseq2 results}
\usage{
read_deseq_tibble(filepath)
}
\arguments{
\item{filepath}{Path to DESeq2 results object form res saved as tsv}
}
\value{
Tibble of DES2 results.
}
\description{
Parse deseq2 results
}
