% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_functions.R
\name{run_msig_enricher}
\alias{run_msig_enricher}
\title{Run the msig using clusterprofiler on a gene set list.}
\usage{
run_msig_enricher(
  gset_list,
  category = NULL,
  species = "Mus musculus",
  subcategory = NULL,
  GSEA = TRUE,
  universe = NULL,
  translation_table = NULL,
  msdb_var = "gene_symbol",
  input_type = "gene_symbol",
  ...
)
}
\arguments{
\item{gset_list}{A data.frame/tibble with enriched genes and their log fold changes}

\item{category}{Msigdb categories to include. Defaults to NULL (all categories)}

\item{species}{Latin name of species where data comes from. Defaults to Mus musculus.}

\item{subcategory}{Msigdb subcategory}

\item{GSEA}{logical argument if GSEA is being used or over representation test}

\item{universe}{universe of all tested genes to be given when GSEA is FALSE}

\item{translation_table}{named list translating gene IDs from \link{table_to_list}}

\item{msdb_var}{Variable to use for gsea from msgdbr table \link[msigdbr]{msigdbr}}

\item{input_type}{Type of input in gset_list}

\item{...}{Parameters passed to enichment functions from clusterProfiler}
}
\value{
\link[DOSE]{enrichResult-class} of given information
}
\description{
Run the msig using clusterprofiler on a gene set list.
}
