% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{save_cheatmap_svg}
\alias{save_cheatmap_svg}
\title{Save pheatmapplot to svg  file}
\usage{
save_cheatmap_svg(x, filename, width = 7, height = 7)
}
\arguments{
\item{x}{plot object to save}

\item{filename}{name where to save file}

\item{width}{plotwidth}

\item{height}{plotheight}
}
\value{
nothing
}
\description{
Save pheatmapplot to svg  file
}
