% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_functions.R
\name{save_dotplots}
\alias{save_dotplots}
\title{Title}
\usage{
save_dotplots(d_plot, p_group, p_path = plot_path, gsea_type)
}
\arguments{
\item{d_plot}{dotplot object}

\item{p_group}{plot_group}

\item{p_path}{plot_path}

\item{gsea_type}{Type of gsea}
}
\description{
Title
}
