% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senes_gs.R
\docType{data}
\name{senes}
\alias{senes}
\title{Senescence Gene sets}
\format{
Ein Dataframe mit x Zeilen und y Spalten:
\describe{
  \item{source}{Name of the gene set the gene in target belongs to}
  \item{target}{Name of the gene as a gene symbol}

}
}
\usage{
senes
}
\description{
This object contians the gene sets for senescence genes.
}
\keyword{datasets}
