% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{split_contrast_groups}
\alias{split_contrast_groups}
\title{Splits string at -vs- to extract the two contrast groups.}
\usage{
split_contrast_groups(cont)
}
\arguments{
\item{cont}{string containing the contrast groups}
}
\value{
A character vector of length two with the contrast split apart
}
\description{
Splits string at -vs- to extract the two contrast groups.
}
