% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_functions.R
\name{table_to_list}
\alias{table_to_list}
\title{Take two columns from a data frame and move these to a list}
\usage{
table_to_list(tb, ref_col, val_col)
}
\arguments{
\item{tb}{Tibble containing the two columns to transform}

\item{ref_col}{Reference column to be used a name}

\item{val_col}{Value Columns containing values}
}
\description{
Useful for dealing with duplicated intries to build a translation table
}
