% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_functions.R
\name{transform_glist}
\alias{transform_glist}
\title{transform gene set using translation table}
\usage{
transform_glist(gene_set, gene2id)
}
\arguments{
\item{gene_set}{two_column gene_set data_frame to transform}

\item{gene2id}{Translation table from \link{table_to_list}}
}
\description{
Translation table bulilt by \link{table_to_list}
}
