% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnival_helpers.R
\name{convertHumanGeneHomologs}
\alias{convertHumanGeneHomologs}
\title{convert between human and mouse homologs using homolog table}
\usage{
convertHumanGeneHomologs(x, jax_database = NULL)
}
\arguments{
\item{x}{List of genes}

\item{jax_database}{Databse from outside to reduce http calls}
}
\description{
convert between human and mouse homologs using homolog table
}
\examples{
convertHumanGeneHomologs(c("APLNR", "APLNR"))
}
