% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_functions.R
\name{get_entrezgenes_from_ensembl}
\alias{get_entrezgenes_from_ensembl}
\title{Get entrezgenenames from ensembl}
\usage{
get_entrezgenes_from_ensembl(gene_names, input_type = "SYMBOL")
}
\arguments{
\item{gene_names}{vecotr of gene symbols to be converted}

\item{input_type}{String denoting the type of input given based on \link[org.Mm.eg.db]{org.Mm.eg.db}}
}
\value{
table of gene names to entrezgene ids
}
\description{
Get entrezgenenames from ensembl
}
\examples{
get_entrezgenes_from_ensembl(c("Aplnr"))
}
