% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnival_helpers.R
\name{rename_count_rownames}
\alias{rename_count_rownames}
\title{rename ENSG IDS from OTP using fpkm table}
\usage{
rename_count_rownames(count_table, fpkm_table)
}
\arguments{
\item{count_table}{Table with ENSG IDs with version as rownames}

\item{fpkm_table}{FPKM table from rna-star-deseq pipeline `fpkm/all.tsv`}
}
\description{
rename ENSG IDS from OTP using fpkm table
}
\examples{
NULL
}
