% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_functions.R
\name{run_gsea_query}
\alias{run_gsea_query}
\title{Function to run all gsea queries based on config file}
\usage{
run_gsea_query(
  gsea_genes,
  de_genes,
  gset_name,
  gset_config,
  species,
  org_db,
  t_table = NULL
)
}
\arguments{
\item{gsea_genes}{tibble with gene ids in first column and gsea statistic in second}

\item{de_genes}{tibble with DE genes in first column}

\item{gset_name}{Name of gene set in config to analyze}

\item{gset_config}{gset config object}

\item{species}{species name}

\item{org_db}{org.mm.eg.db}

\item{t_table}{translation_table to convert EnsemblIDs to gene symbols}
}
\value{
List with results enrichment analysis \link[DOSE]{enrichResult-class}
}
\description{
Function to run all gsea queries based on config file
}
